refactor 24 hours stats
