+++ stats/stats/src/charts/counters/mod.rs
@@ -1,9 +1,7 @@
 mod average_block_time;
-mod average_txn_fee_24h;
 mod completed_txns;
 mod last_new_contracts;
 mod last_new_verified_contracts;
-mod new_txns_24h;
 mod pending_txns;
 mod total_accounts;
 mod total_addresses;
@@ -15,19 +13,16 @@ mod total_operational_txns;
 mod total_tokens;
 mod total_txns;
 mod total_verified_contracts;
-mod txns_fee_24h;
+mod txns_stats_24h;
 mod yesterday_txns;
 
 #[cfg(test)]
 mod mock;
 
 pub use average_block_time::AverageBlockTime;
-pub use average_txn_fee_24h::AverageTxnFee24h;
-pub(crate) use average_txn_fee_24h::TxnsStatsValue;
 pub use completed_txns::CompletedTxns;
 pub use last_new_contracts::LastNewContracts;
 pub use last_new_verified_contracts::LastNewVerifiedContracts;
-pub use new_txns_24h::NewTxns24h;
 pub use pending_txns::PendingTxns;
 pub use total_accounts::TotalAccounts;
 pub use total_addresses::TotalAddresses;
@@ -39,7 +34,10 @@ pub use total_operational_txns::TotalOperationalTxns;
 pub use total_tokens::TotalTokens;
 pub use total_txns::{TotalTxns, TotalTxnsInt};
 pub use total_verified_contracts::TotalVerifiedContracts;
-pub use txns_fee_24h::TxnsFee24h;
+pub(crate) use txns_stats_24h::TxnsStatsValue;
+pub use txns_stats_24h::{
+    average_txn_fee_24h::AverageTxnFee24h, new_txns_24h::NewTxns24h, txns_fee_24h::TxnsFee24h,
+};
 pub use yesterday_txns::YesterdayTxns;
 
 #[cfg(test)]
