+++ stats/stats/src/charts/counters/txns_stats_24h/average_txn_fee_24h.rs
@@ -5,35 +5,35 @@ use crate::{
     },
     gettable_const,
     types::TimespanValue,
-    ChartError, ChartProperties, MissingDatePolicy, Named,
+    ChartProperties, MissingDatePolicy, Named,
 };
 use chrono::NaiveDate;
 use entity::sea_orm_active_enums::ChartType;
 
-use super::{average_txn_fee_24h::Txns24hStats, TxnsStatsValue};
+use super::{Txns24hStats, TxnsStatsValue};
 
-pub struct ExtractSum;
+pub struct ExtractAverage;
 
-impl MapFunction<TimespanValue<NaiveDate, TxnsStatsValue>> for ExtractSum {
+impl MapFunction<TimespanValue<NaiveDate, TxnsStatsValue>> for ExtractAverage {
     type Output = TimespanValue<NaiveDate, Option<f64>>;
 
     fn function(
         inner_data: TimespanValue<NaiveDate, TxnsStatsValue>,
-    ) -> Result<Self::Output, ChartError> {
+    ) -> Result<Self::Output, crate::ChartError> {
         Ok(TimespanValue {
             timespan: inner_data.timespan,
-            value: inner_data.value.fee_sum,
+            value: inner_data.value.fee_average,
         })
     }
 }
 
-pub type TxnsFee24hExtracted = Map<Txns24hStats, ExtractSum>;
+pub type AverageTxnFee24hExtracted = Map<Txns24hStats, ExtractAverage>;
 
 pub struct Properties;
 
 impl Named for Properties {
     fn name() -> String {
-        "txnsFee24h".into()
+        "averageTxnFee24h".into()
     }
 }
 
@@ -44,14 +44,16 @@ impl ChartProperties for Properties {
         ChartType::Counter
     }
     fn missing_date_policy() -> MissingDatePolicy {
-        MissingDatePolicy::FillPrevious
+        MissingDatePolicy::FillZero
     }
 }
 
 gettable_const!(Zero: f64 = 0.0);
 
-pub type TxnsFee24h =
-    DirectPointLocalDbChartSource<MapToString<UnwrapOr<TxnsFee24hExtracted, Zero>>, Properties>;
+pub type AverageTxnFee24h = DirectPointLocalDbChartSource<
+    MapToString<UnwrapOr<AverageTxnFee24hExtracted, Zero>>,
+    Properties,
+>;
 
 #[cfg(test)]
 mod tests {
@@ -60,28 +62,33 @@ mod tests {
 
     #[tokio::test]
     #[ignore = "needs database to run"]
-    async fn update_txns_fee_24h_1() {
-        simple_test_counter::<TxnsFee24h>("update_txns_fee_24h_1", "0.000023592592569", None).await;
-    }
-
-    #[tokio::test]
-    #[ignore = "needs database to run"]
-    async fn update_txns_fee_24h_2() {
-        simple_test_counter::<TxnsFee24h>(
-            "update_txns_fee_24h_2",
-            "0.000495444443949",
-            Some(dt("2022-11-11T00:00:00")),
+    async fn update_average_txns_fee_1() {
+        simple_test_counter::<AverageTxnFee24h>(
+            "update_average_txns_fee_1",
+            "0.000023592592569",
+            None,
         )
         .await;
     }
 
     #[tokio::test]
     #[ignore = "needs database to run"]
-    async fn update_txns_fee_24h_3() {
-        simple_test_counter::<TxnsFee24h>(
-            "update_txns_fee_24h_3",
+    async fn update_average_txns_fee_2() {
+        simple_test_counter::<AverageTxnFee24h>(
+            "update_average_txns_fee_2",
+            "0.00006938997814411765",
+            Some(dt("2022-11-11T16:00:00")),
+        )
+        .await;
+    }
+
+    #[tokio::test]
+    #[ignore = "needs database to run"]
+    async fn update_average_txns_fee_3() {
+        simple_test_counter::<AverageTxnFee24h>(
+            "update_average_txns_fee_3",
             "0",
-            Some(dt("2024-11-11T00:00:00")),
+            Some(dt("2024-10-10T00:00:00")),
         )
         .await;
     }
